this code is currently unused
git-svn-id: http://svn.vuze.com/public/client@8177 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/com/aelitis/azureus/core/peermanager/piecepicker/priority/impl/PriorityShapeImpl.java
@@ -31,7 +31,7 @@ abstract public class PriorityShapeImpl
 	implements PriorityShape
 {
 	public int	mode =0;
-	public int	priority =0;	
+	public long	priority =0;	
 
 	public PriorityShapeImpl()
 	{
@@ -49,7 +49,7 @@ abstract public class PriorityShapeImpl
 	}
 
 	
-	public int getPriority()
+	public long getPriority()
 	{
 		return priority;
 	}
